RelationAPI
###작업내용###
- 멤버리스트 조회를 page방식으로 수정
- Relation API 기능별 완성
- Relation State 에 FOLLOWED 추가 : 수동적 의미를 표현하기 위해 추가했고 이 state는 database에 직접적으로 저장되지 않으며 리스트 조회시에 다른 function에 접근하기 위해 사용함

+++ src/main/java/com/d2/timeline/domain/service/RelationService.java
@@ -4,22 +4,28 @@ import com.d2.timeline.domain.common.RelationState;
 import com.d2.timeline.domain.dao.MemberRepository;
 import com.d2.timeline.domain.dao.UserRelationRepository;
 import com.d2.timeline.domain.dto.MemberDTO;
+import com.d2.timeline.domain.dto.UserRelationDTO;
 import com.d2.timeline.domain.vo.Member;
 import com.d2.timeline.domain.vo.UserRelation;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
+import org.springframework.stereotype.Service;
 
-import java.util.Collections;
-import java.util.List;
-import java.util.ArrayList;
-import java.util.Optional;
+import javax.persistence.NoResultException;
+import java.util.*;
 
 import static com.d2.timeline.domain.Constant.RelationServiceConstant.*;
 
+@Service
 public class RelationService {
 
     private static final Logger logger = LoggerFactory.getLogger(MemberFindService.class);
+
 
     @Autowired
     MemberRepository memberRepository;
@@ -28,34 +34,38 @@ public class RelationService {
     UserRelationRepository userRelationRepo;
 
     public UserRelation createRelation(Member master, Member slave, RelationState state){
-        UserRelation userRelation = UserRelation.builder()
-                .master(master)
-                .slave(slave)
-                .state(state)
-                .build();
 
+        Optional<UserRelation> maybeUserRelation = userRelationRepo.findByMasterAndSlave(master, slave);
+        UserRelation userRelation = maybeUserRelation.orElseGet(()->{
+           return UserRelation.builder()
+                   .master(master)
+                   .slave(slave)
+                   .state(state)
+                   .build();
+        });
+
+        userRelation.setState(state);
+        userRelationRepo.save(userRelation);
         return userRelation;
     }
 
     public String deleteRelation(Member master, Member slave){
+
         boolean exist = userRelationRepo.existsByMasterAndSlave(master, slave);
 
         if(!exist) {
-            return "NOT EXIST";
+            return ERROR_NOT_EXIST;
         }
 
         userRelationRepo.deleteByMasterAndSlave(master, slave);
-        return "OK";
+        return OK_MSG_RELATION;
     }
 
     public String updateState(Member master, Member slave, RelationState state){
 
         Optional<UserRelation> maybeUserRelation = userRelationRepo.findByMasterAndSlave(master, slave);
-        UserRelation userRelation  = maybeUserRelation.orElseGet(()->createRelation(master, slave, state));
-
-        if(userRelation.getId() == -1){
-            return STATE_CHANGE_ERROR_MSG;
-        }
+        UserRelation userRelation  = maybeUserRelation.orElseGet(
+                ()-> createRelation(master, slave, state));
 
         userRelation.setState(state);
         userRelationRepo.save(userRelation);
@@ -63,12 +73,27 @@ public class RelationService {
         return STATE_CHANGE_OK_MSG;
     }
 
-    public String followRequest(Member master, Member slave){
+    public String followRequest(UserRelationDTO userRelationDTO){
+        Member master = (memberRepository.findById(userRelationDTO.getMasterId())).orElseThrow(
+                ()-> new NoResultException(ERROR_NOT_EXIST));
+        Member slave = (memberRepository.findById(userRelationDTO.getSlaveId())).orElseThrow(
+                ()-> new NoResultException(ERROR_NOT_EXIST));
+
         createRelation(master, slave, RelationState.REQUEST);
         return FOLLOW_REQUEST_OK_MSG;
     }
 
-    public String responseForFollowRequest(Member requester, Member respondent, boolean allow){
+    public String responseForFollowRequest(UserRelationDTO userRelationDTO, boolean allow){
+        Member requester = (memberRepository.findById(userRelationDTO.getMasterId())).orElseThrow(
+                ()-> new NoResultException(ERROR_NOT_EXIST));
+        Member respondent = (memberRepository.findById(userRelationDTO.getSlaveId())).orElseThrow(
+                ()-> new NoResultException(ERROR_NOT_EXIST));
+
+        boolean exist = userRelationRepo.existsByMasterAndSlaveAndState(requester, respondent, RelationState.REQUEST);
+
+        if(!exist) {
+            return ERROR_NOT_EXIST;
+        }
 
         if(!allow){
             deleteRelation(requester, respondent);
@@ -80,47 +105,79 @@ public class RelationService {
         return FOLLOW_RESPONSE_OK_MSG;
     }
 
-    public String block(Member master, Member slave){
+    public String unfollow(UserRelationDTO userRelationDTO){
+        Member master = (memberRepository.findById(userRelationDTO.getMasterId())).orElseThrow(
+                ()-> new NoResultException(ERROR_NOT_EXIST));
+        Member slave = (memberRepository.findById(userRelationDTO.getSlaveId())).orElseThrow(
+                ()-> new NoResultException(ERROR_NOT_EXIST));
+        UserRelation userRelation = userRelationRepo.findByMasterAndSlave(master, slave).orElseThrow(
+                ()-> new NoResultException(ERROR_NOT_EXIST));
+
+        if(userRelation.getState() != RelationState.FOLLOW){
+            return ERROR_BAD_REQUEST;
+        }
+
+        deleteRelation(master, slave);
+        return OK_MSG_RELATION;
+    }
+
+    public String block(UserRelationDTO userRelationDTO){
+
+        Member master = (memberRepository.findById(userRelationDTO.getMasterId())).orElseThrow(
+                ()-> new NoResultException(ERROR_NOT_EXIST));
+        Member slave = (memberRepository.findById(userRelationDTO.getSlaveId())).orElseThrow(
+                ()-> new NoResultException(ERROR_NOT_EXIST));
 
         updateState(master, slave, RelationState.BLOCK);
         updateState(slave, master, RelationState.BLOCKED);
-        return "OK";
+        return OK_MSG_RELATION;
 
     }
 
-    public String unblock(Member master, Member slave){
+    public String unblock(UserRelationDTO userRelationDTO){
+
+        Member master = (memberRepository.findById(userRelationDTO.getMasterId())).orElseThrow(
+                ()-> new NoResultException(ERROR_NOT_EXIST));
+        Member slave = (memberRepository.findById(userRelationDTO.getSlaveId())).orElseThrow(
+                ()-> new NoResultException(ERROR_NOT_EXIST));
+        UserRelation userRelation = userRelationRepo.findByMasterAndSlave(master, slave).orElseThrow(
+                ()-> new NoResultException(ERROR_NOT_EXIST));
+        UserRelation reverseRelation = userRelationRepo.findByMasterAndSlave(slave, master).orElseThrow(
+                ()-> new NoResultException(ERROR_NOT_EXIST));
+
+        boolean canUnblockState = (userRelation.getState() == RelationState.BLOCK)
+                && (reverseRelation.getState() == RelationState.BLOCKED);
+        if(!canUnblockState){
+            return ERROR_BAD_REQUEST;
+        }
 
         deleteRelation(master, slave);
         deleteRelation(slave, master);
 
-        return "OK";
+        return OK_MSG_RELATION;
     }
 
-    public List<MemberDTO> showRelationList(Member member, RelationState state){
 
-        List<MemberDTO> memberDTOList = new ArrayList<>();
-        List<Member> relationMemberList = userRelationRepo.findSlaveByMasterAndState(member, state).orElseGet(()->{
-            logger.info(RELATION_NO_EXIST_MSG);
-            return Collections.emptyList();
-        });
+    public Page<MemberDTO> showRelationList(Long memberId, RelationState state, Pageable pageable){
+        Member member = memberRepository.findById(memberId).orElseThrow(
+                ()-> new NoResultException(ERROR_NOT_EXIST));
+        if(state == RelationState.FOLLOWED){
+            return showFollowerList(memberId, pageable);
+        }
+        Page<Member> relationMemberList = userRelationRepo.findSlaveByMasterAndState(member, state, pageable);
 
-        relationMemberList.forEach(x -> memberDTOList.add(new MemberDTO(x)));
-
-        return memberDTOList;
+        return relationMemberList.map(MemberDTO::new);
     }
 
-    public List<MemberDTO> showFollowerList(Member member){
+    public Page<MemberDTO> showFollowerList(Long memberId, Pageable pageable){
+        RelationState state = RelationState.FOLLOW;
+        Member member = memberRepository.findById(memberId).orElseThrow(
+                ()-> new NoResultException(ERROR_NOT_EXIST));
+        Page<Member> relationMemberList = userRelationRepo.findMasterBySlaveAndState(member, state, pageable);
 
-        List<MemberDTO> memberDTOList = new ArrayList<>();
-        List<Member> relationMemberList = userRelationRepo.findMasterBySlaveAndState(member, RelationState.FOLLOW).orElseGet(()->{
-            logger.info(RELATION_NO_EXIST_MSG);
-            return Collections.emptyList();
-        });
-
-        relationMemberList.forEach(x -> memberDTOList.add(new MemberDTO(x)));
-
-        return memberDTOList;
+        return relationMemberList.map(MemberDTO::new);
     }
+
